Switched back float DPs to double.

+++ Microsoft.Toolkit.Uwp.UI.Controls/ImageCropper/ImageCropper.Events.cs
@@ -137,10 +137,10 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
         {
             var imageCropperThumb = (ImageCropperThumb)sender;
             var currentPointerPosition = new Point(
-                imageCropperThumb.X + e.Position.X + e.Delta.Translation.X - ((float)imageCropperThumb.ActualWidth / 2),
-                imageCropperThumb.Y + e.Position.Y + e.Delta.Translation.Y - ((float)imageCropperThumb.ActualHeight / 2));
+                (float)imageCropperThumb.X + e.Position.X + e.Delta.Translation.X - ((float)imageCropperThumb.ActualWidth / 2),
+                (float)imageCropperThumb.Y + e.Position.Y + e.Delta.Translation.Y - ((float)imageCropperThumb.ActualHeight / 2));
             var safePosition = GetSafePoint(_restrictedSelectRect, currentPointerPosition);
-            var safeDiffPoint = new Point(safePosition.X - imageCropperThumb.X, safePosition.Y - imageCropperThumb.Y);
+            var safeDiffPoint = new Point(safePosition.X - (float)imageCropperThumb.X, safePosition.Y - (float)imageCropperThumb.Y);
             UpdateCroppedRect(imageCropperThumb.Position, safeDiffPoint);
         }
